requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
